Ajustando testes unitários

+++ src/AlbionOnlineSniffer.Core/Handlers/HealthUpdateEventHandler.cs
@@ -21,8 +21,8 @@ namespace AlbionOnlineSniffer.Core.Handlers
 
         protected override async Task OnActionAsync(HealthUpdateEvent value)
         {
-            playerHandler.UpdateHealth(value.Id, value.Health);
-            mobHandler.UpdateHealth(value.Id, value.Health);
+            playerHandler.UpdateHealth(value.Id, (int)value.Health);
+            mobHandler.UpdateHealth(value.Id, (int)value.Health);
             
             // Emitir evento para o EventDispatcher
             await eventDispatcher.DispatchEvent(value);
